Rewrote the core merging code. While it is now somewhat less optimized, and might be writing to the one or other string buffer more than it absolutely would have to, it also is 500% more readable, understandable an extensible. I hope I learned my lesson here. KISS!
In the process, the filter API changed again, in an
backwards-incompatible way. As was already noted in the TODO, the
previous separation between target and source filters was crap. Now,
a filter can both act as a source and target filter at the same time,
implementing input() and output() methods (inspired by
django_compressor).

Also part of this commit are a bunch of new tests; the actual
building/merging/filtering process is now tested for the first time.