Clean up the code we use to determine cache keys.
Move the md5 calcuation of the key for a hunk into the BaseHunk class.

The cache API now supports any type of key that is hash()able in Python.
Depending on the type of cache a class implements, it may or may not
need to call hash() to determine a key, or can work directly with the
Python objects.

We no longer need to make_key() function which ways always kind of
confusing in it's purpose, as evidenced by the nested call to it that we
had.