Properly deal with the "debug" property of child bundles.
So far, the debug of a child bundle only worked in a certain narrow set
of circumstances: Only when globally in debug mode, and only for the
first bundle that declares it's own debug option.

Now, all combinations should be properly handled, and invalid
combinations (a child bundle trying to enter debug mode) should properly
raise errors.

The tests related to this received some attention, and should now
provide better coverage (a number of scenarios are tested that weren't
before, in particular the merge mode), and be slightly more to the point
and better designed.