Added Environment._normalize_source_path() for extensions to override.
This will allow extensions like django-assets to implement virtual
file resolving like required by django.contrib.staticfiles.

The new method is quasi-official, as opposed to overwriting
Environment.abspath(), which was used for example by Flask-Assets
up to this point. The change is backwards-incompatbile in that
respect - Flask-Assets will need to be updated.

Further, Bundle.resolve_contents() now takes care of converting
everything to absolute paths, and raises an error early if a source
file does not exist. This means that invalid source files will
now raise an error even in debug mode.