No longer raise a BuildError when updater=False and output file missing.
An exception would be raised that with env.updater=False, a Bundle's output file would not be autocreated. However, the convenience of this exception is outweighed by it's downsides:

- A filesystem stat request needs to be issued (at least with the timestamp-based updater), which costs performance. For a user with env.updater disabled, who wants to do builds herself, this is a waste.

- A user who disables env.updater is also more likely have a deployment scenario where the media files are deployed to a different server. As a result, we aren't even able to access the output file, and are thus causing the user issues. An example for this is Google App Engine.

In exchange, the API of Bundle.build() has been finetuned so that without an updater configured, the 'force' argument defaults to True.

Relates to #74.