Pick up changes in dependencies when force=True.
The issue was that a change in a bundle dependency (i.e. for example a sass include file) did not cause the cache key to change, and thus the build using the cache and losing the change.

A similar issue was fixed previously, but that fix was applied incorrectly at the level of the updater, so that while it worked in development, it didn't in deployments which don't use an updater, but instead rely on the "rebuild" command, for example.