Consider url sources when outputting urls.
While this is the right thing to do, it worked fine without this already. However,
third party extensions which overwrite absurl() had then to pay attention and
support url sources themselves, which at least Flask-Assets did not. Which is how
this turned up.

Now, bundle._make_url() and env.absurl() will not be called for url sources.