Make the setup() call part of the filter protocol.
setup() will now always be called before a filter is used with a bundle
(as opposed to only when it is used with a new environment).

As a result, the JST filter can now properly work (fixes #104).

There should be no performance concerns here. What filters do in
setup() should generally be really quick, and since already multiple
uses of a filter generally imply multiple instances, not much should
change in terms of overhead.