Rename the first() filter to open().
Change it's API to not receive a pre-opened file. This is a lot
clearer, and good for performance too.

Caching needed some touching, because filter operations like open()
can now have arbitrary arguments, and a proper cache key needs to
be created.

And God knows why the MemoryCache class was duplicated.