Fixed input() kwargs when running after open().
There was a bug where an input() filter would not receive the source_path kwarg
if an open() filter had been run before. This was because source_path was taken
strictly from the FileHunk.filename attribute, breaking when a previous filter
cause a MemoryHunk to be used.

One problem caused by this was the cssrewrite filter being broken.