Limited ability for filters to run in debug mode.
Filters can now declare that they need to run in 'merge' mode (cssrewrite),
or that they need to run always (compiler-type filters).

In the latter case, if such a filter is used, the bundle will automatically be
merged. This is a more limited change than it sounds - still, see the section
on backwards-incompatibilities in the docs for more info.

Closes #60, and solves the requests in #73, #102 and possibly others.
Some relation to #120.