Subtly change auto_build behavior.
It used to be the case that bundle.build(force=False) with auto_build=False did
nothing at all, which is stupid. The problem was that the way auto_build=False
caused the build() ``force`` argument to default to True (in order for such an
call to not also be a noop) was fundamentally flawed.

Instead, the correct way to think about this is that auto_build only affects
building during an urls() call, and manually calling build() does what the name
says: it builds.

This new approach is indefinitely better; it makes sense and is actually
simpler, as evidenced by the net code-removal of this commit.