Make Bundle.urls() pull external files.
If a proper url to a file cannot be generated due to it not being within
Environment.directory, copy the file into the directory.

This is actually a pretty neat feature during development: You can for example
reference a develoment version of an external library that you are also working
on in a separate project.