Bundle.build(): disable_cache=False behavior.
This used to override the auto-select behavior. The CLI interface used this,
as a result of which the cache was not disabled when running a build command,
in cases where it should have.

This could also have been changed in the CLI code, but it's an easy mistake to
make, and there really is no good reason to use the cache in cases where it
clearly can lead to incorrect results.