Introduce 'prefix' option for bundles.
A prefix makes it possible to redirect all or some of the static files to do some post-processing on the files. Such post-processing might be using a CSS compiler, linting, reformatting etc.

The idea is to aid development. Currently, webassets merges static files if a CSS/JS compiler is used, even in development, however, unmerged files are best suited for debugging. At the same time, compilation might be needed even in development, e.g. when some of the bundle's contents are LESS files and browser-based less.js is unsuited. Same is probably also true for Django templates.

So now, with debugging option on the bundle, files are returned as is, which is unsuitable for developing with LESS. Without it, files are combined, which is fine for production, but unsuitable for development. With the prefix option and debugging on, pseudo-bundle processing can be done in a custom view and thus be suitable for development with tools like LESS.

An example of such post-processing view for Django can be found at https://gist.github.com/3206764