Ported cache system to Python 3.
I changed the API for this; before, the filesystem cache tried to
write strings directly, and only pickled other types, and required
some help from the caller on get() to know whether to unpickle.

Apart from being confusing, this was also a bit of a mindfuck trying
to make it work both on Python 2 and 3 (which is not to say it wouldn't
have been possible).

I've changed the BaseCache API a bit, and the filesystem cache now
simply pickles everything.