Remove gzip filter.
It turns out that it is a problem on Python 2 as well. Obviously,
merging the output of the gzip filter with that of a text filter
(which all use unicode) wouldn't work. While that makes little
sense in practice, it turns out that the code that processes
filters failed with the gzip output due to it's assumptions about
unicode. While I could have worked around this in theory, I don't
think it's worthwhile adding another hack for this.

Either we want gzip or not, and if we do, we'll need to find a
proper solution that works on Python 3 as well.