Cleancss as an input filter to fix import inlining
cleancss is the only webassets filter(AFAIK) that supports CSS import inlining. But to do that it needs to know the directory path of the CSS file it's working on, which is only accessible to input filters.

We provide an input function that passes the --root argument to cleancss to fix this issue.

We keep the output filter as an added bonus to remove new lines between files processed this way.