Proof of concept support for calculating SRI for jinja2 backend.
This is the start of the implementation of #494

Currently only tested on Python 3.7, missing unit tests and documentation.

Basic idea is to extend the Bundle.urls() function to return a list of dictionaries {'uri': <same as before>, 'sri': <hash>} but only when requested (to not break existing users). A dict is used instead of a tuple to allow adding other values in the future. Currently only works for local resources.