Use the safe load method in the YAML loader to avoid a warning
The YAML loader is currently using the `yaml.load()` method without
defining a loader, which is deprecated [1] and causes a warning to
appear:

```
YAMLLoadWarning: calling yaml.load() without Loader=... is deprecated, as the default Loader is unsafe. Please read https://msg.pyyaml.org/load for full details.
  obj = self.yaml.load(f) or {}
```

This patch updates the call to use `yaml.safe_load()` which is a
shortcut method to use the `SafeLoader` class.

Fixes #517

[1] https://github.com/yaml/pyyaml/wiki/PyYAML-yaml.load(input)-Deprecation